refactor: Apple OAuth 패키지 구조 정리 및 네이밍 통일
