+++ love/infrastructure/src/main/kotlin/com/yapp/love/infrastructure/oauth/apple/client/AppleClientSecretService.kt
@@ -1,30 +1,26 @@
-package com.yapp.love.infrastructure.oauth.apple.service
+package com.yapp.love.infrastructure.oauth.apple.client
 
 import com.nimbusds.jose.JWSAlgorithm
 import com.nimbusds.jose.JWSHeader
-import com.yapp.love.infrastructure.oauth.apple.ApplePrivateKeyLoader
-import com.yapp.love.infrastructure.oauth.apple.config.AppleOauthProperties
+import com.nimbusds.jose.crypto.ECDSASigner
+import com.nimbusds.jwt.JWTClaimsSet
+import com.nimbusds.jwt.SignedJWT
+import com.yapp.love.infrastructure.oauth.apple.config.AppleOAuthProperties
 import org.springframework.stereotype.Service
 import java.time.Instant
 import java.time.temporal.ChronoUnit
 import java.util.Date
 
-/**
- * Apple client_secret JWT 생성 서비스
- *
- * Apple /auth/token API 호출 시 필요한 client_secret을 생성합니다.
- * ES256 알고리즘으로 서명된 JWT 형식이며, 유효시간은 5분입니다.
- */
 @Service
 class AppleClientSecretService(
-    private val appleProperties: AppleOauthProperties,
+    private val appleProperties: AppleOAuthProperties,
     private val keyLoader: ApplePrivateKeyLoader,
 ) {
     fun createClientSecret(now: Instant = Instant.now()): String {
         val exp = now.plus(5, ChronoUnit.MINUTES)
 
         val claims =
-            com.nimbusds.jwt.JWTClaimsSet.Builder()
+            JWTClaimsSet.Builder()
                 .issuer(appleProperties.teamId)
                 .subject(appleProperties.clientId)
                 .audience(appleProperties.aud)
@@ -37,8 +33,8 @@ class AppleClientSecretService(
                 .keyID(appleProperties.keyId)
                 .build()
 
-        val signedJwt = com.nimbusds.jwt.SignedJWT(header, claims)
-        val signer = com.nimbusds.jose.crypto.ECDSASigner(keyLoader.privateKey)
+        val signedJwt = SignedJWT(header, claims)
+        val signer = ECDSASigner(keyLoader.privateKey)
 
         signedJwt.sign(signer)
         return signedJwt.serialize()
