+++ love/infrastructure/src/main/kotlin/com/yapp/love/infrastructure/oauth/apple/client/AppleTokenResponse.kt
@@ -1,0 +1,13 @@
+package com.yapp.love.infrastructure.oauth.apple.client
+
+import com.fasterxml.jackson.databind.PropertyNamingStrategies
+import com.fasterxml.jackson.databind.annotation.JsonNaming
+
+@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy::class)
+data class AppleTokenResponse(
+    val accessToken: String,
+    val idToken: String,
+    val refreshToken: String?,
+    val expiresIn: Int,
+    val tokenType: String,
+)
