+++ love/infrastructure/src/main/kotlin/com/yapp/love/infrastructure/oauth/apple/config/AppleOauthProperties.kt
@@ -3,7 +3,7 @@ package com.yapp.love.infrastructure.oauth.apple.config
 import org.springframework.boot.context.properties.ConfigurationProperties
 
 @ConfigurationProperties(prefix = "apple.oauth")
-data class AppleOauthProperties(
+data class AppleOAuthProperties(
     val clientId: String,
     val teamId: String,
     val keyId: String,
