--- love/infrastructure/src/main/kotlin/com/yapp/love/infrastructure/oauth/apple/response/AppleTokenResponse.kt
@@ -1,16 +1,0 @@
-package com.yapp.love.infrastructure.oauth.apple.response
-
-import com.fasterxml.jackson.annotation.JsonProperty
-
-data class AppleTokenResponse(
-    @JsonProperty("access_token")
-    val accessToken: String,
-    @JsonProperty("id_token")
-    val idToken: String,
-    @JsonProperty("refresh_token")
-    val refreshToken: String?,
-    @JsonProperty("expires_in")
-    val expiresIn: Int,
-    @JsonProperty("token_type")
-    val tokenType: String,
-)
